@@ -2,13 +2,12 @@ apply plugin: 'com.android.application'
 apply plugin: 'com.diffplug.spotless'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.2"
+    compileSdkVersion 33
 
     defaultConfig {
         applicationId "br.net.du.arhaticyogajournal"
         minSdkVersion 14
-        targetSdkVersion 29
+        targetSdkVersion 33
     }
 
     buildTypes {
@@ -23,13 +22,21 @@ dependencies {
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation 'com.github.clans:fab:1.6.4'
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:1.9.5'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
+    testImplementation 'org.mockito:mockito-inline:4.6.1'
 }
 
-configurations.compile {
+configurations.implementation {
     spotless {
         java {
+            // Ref.: https://github.com/diffplug/spotless/issues/240#issuecomment-385206606
+            custom 'noWildcardImports', {
+                if (it.contains('*;\n')) {
+                    throw new Error('No wildcard imports allowed')
+                }
+            }
+            bumpThisNumberIfACustomStepChanges(1)
+
             googleJavaFormat().aosp()
             target '**/*.java'
         }
